Fixed resize adjustment and added some generic resolutions.

+++ src/model/MainFood.java
@@ -5,7 +5,7 @@ public class MainFood {
     public int x;
     public int y;
 
-    public void updateLocation(int factorX, int factorY) {
+    public void updateLocation(float factorX, float factorY) {
         x *= factorX;
         y *= factorY;
     }
