+++ src/model/Snake.java
@@ -8,7 +8,7 @@ public class Snake {
     public int foodEaten;
     public char direction;
 
-    public void updateLocation(int factorX, int factorY) {
+    public void updateLocation(float factorX, float factorY) {
         for (int i = 0; i < bodyParts; i++) {
             x[i] *= factorX;
             y[i] *= factorY;
